Linux/GCC compability

+++ Source/Oscilloscope/ChannelData.h
@@ -80,7 +80,7 @@
 
 				void resizeStorage(std::size_t samples, std::size_t capacity = -1)
 				{
-					if (capacity == -1)
+					if (capacity == static_cast<std::size_t>(-1))
 						capacity = cpl::Math::nextPow2Inc(samples);
 
 					for (auto & c : channels)
